refactor: streamline project service with better reusable functions
