+++ backend/pkg/projects/pull_policy_test.go
@@ -1,0 +1,130 @@
+package projects
+
+import (
+	"testing"
+
+	composetypes "github.com/compose-spec/compose-go/v2/types"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestResolveServiceImagePullMode(t *testing.T) {
+	tests := []struct {
+		name     string
+		service  composetypes.ServiceConfig
+		expected ImagePullMode
+	}{
+		{
+			name:     "default policy is missing",
+			service:  composetypes.ServiceConfig{},
+			expected: ImagePullModeIfMissing,
+		},
+		{
+			name:     "always policy",
+			service:  composetypes.ServiceConfig{PullPolicy: composetypes.PullPolicyAlways},
+			expected: ImagePullModeAlways,
+		},
+		{
+			name:     "refresh policy",
+			service:  composetypes.ServiceConfig{PullPolicy: composetypes.PullPolicyRefresh},
+			expected: ImagePullModeAlways,
+		},
+		{
+			name:     "missing policy",
+			service:  composetypes.ServiceConfig{PullPolicy: composetypes.PullPolicyMissing},
+			expected: ImagePullModeIfMissing,
+		},
+		{
+			name:     "if not present policy",
+			service:  composetypes.ServiceConfig{PullPolicy: composetypes.PullPolicyIfNotPresent},
+			expected: ImagePullModeIfMissing,
+		},
+		{
+			name:     "never policy",
+			service:  composetypes.ServiceConfig{PullPolicy: composetypes.PullPolicyNever},
+			expected: ImagePullModeNever,
+		},
+		{
+			name:     "invalid policy defaults to missing behavior",
+			service:  composetypes.ServiceConfig{PullPolicy: "definitely_invalid"},
+			expected: ImagePullModeIfMissing,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			assert.Equal(t, tt.expected, ResolveServiceImagePullMode(tt.service))
+		})
+	}
+}
+
+func TestBuildProjectImagePullPlan(t *testing.T) {
+	services := composetypes.Services{
+		"web": {
+			Name:       "web",
+			Image:      "redis:latest",
+			PullPolicy: composetypes.PullPolicyIfNotPresent,
+		},
+		"worker": {
+			Name:       "worker",
+			Image:      "redis:latest",
+			PullPolicy: composetypes.PullPolicyAlways,
+		},
+		"api": {
+			Name:       "api",
+			Image:      "nginx:latest",
+			PullPolicy: composetypes.PullPolicyNever,
+		},
+		"empty-image": {
+			Name:       "empty-image",
+			Image:      "",
+			PullPolicy: composetypes.PullPolicyAlways,
+		},
+	}
+
+	plan := BuildImagePullPlan(services)
+
+	assert.Len(t, plan, 2)
+	assert.Equal(t, ImagePullModeAlways, plan["redis:latest"])
+	assert.Equal(t, ImagePullModeNever, plan["nginx:latest"])
+}
+
+func TestNormalizePullPolicy(t *testing.T) {
+	assert.Equal(t, "missing", NormalizePullPolicy("if_not_present"))
+	assert.Equal(t, "build", NormalizePullPolicy(" BUILD "))
+	assert.Equal(t, "", NormalizePullPolicy(""))
+}
+
+func TestDecideDeployImageAction(t *testing.T) {
+	t.Run("build service with explicit build policy", func(t *testing.T) {
+		svc := composetypes.ServiceConfig{
+			PullPolicy: "build",
+			Build:      &composetypes.BuildConfig{Context: "."},
+		}
+
+		decision := DecideDeployImageAction(svc, "")
+		assert.True(t, decision.Build)
+		assert.False(t, decision.PullAlways)
+	})
+
+	t.Run("build service default policy uses pull then fallback build", func(t *testing.T) {
+		svc := composetypes.ServiceConfig{Build: &composetypes.BuildConfig{Context: "."}}
+		decision := DecideDeployImageAction(svc, "")
+		assert.True(t, decision.PullIfMissing)
+		assert.True(t, decision.FallbackBuildOnPullFail)
+		assert.False(t, decision.Build)
+	})
+
+	t.Run("non-build service never policy requires local only", func(t *testing.T) {
+		svc := composetypes.ServiceConfig{PullPolicy: "never"}
+		decision := DecideDeployImageAction(svc, "")
+		assert.True(t, decision.RequireLocalOnly)
+		assert.False(t, decision.PullIfMissing)
+	})
+
+	t.Run("compose pull policy wins over deploy override", func(t *testing.T) {
+		svc := composetypes.ServiceConfig{PullPolicy: "never"}
+		decision := DecideDeployImageAction(svc, "always")
+		assert.True(t, decision.RequireLocalOnly)
+		assert.False(t, decision.PullAlways)
+	})
+}
